+++ Runtime/Jobs/AJobBehaviour.cs
@@ -1,3 +1,4 @@
+using Gilzoide.UpdateManager.Jobs.Internal;
 using UnityEngine;
 
 namespace Gilzoide.UpdateManager.Jobs
@@ -12,9 +13,19 @@ namespace Gilzoide.UpdateManager.Jobs
     public abstract class AJobBehaviour<TData> : MonoBehaviour, ITransformJobUpdatable<TData>
         where TData : struct, IUpdateTransformJob
     {
+        /// <summary>
+        /// Whether job data should be synchronized every frame.
+        /// By default, returns false.
+        /// </summary>
+        /// <remarks>
+        /// Override in subclasses implementing <see cref="IJobDataSynchronizer{}"/> and return true
+        /// to automatically register objects for job data synchronization every frame.
+        /// </remarks>
+        public virtual bool SynchronizeJobDataEveryFrame => false;
+
         protected virtual void OnEnable()
         {
-            this.RegisterInManager();
+            this.RegisterInManager(SynchronizeJobDataEveryFrame);
         }
 
         protected virtual void OnDisable()
@@ -38,11 +49,34 @@ namespace Gilzoide.UpdateManager.Jobs
         /// </summary>
         /// <seealso cref="UpdateTransformJobManager{}.GetData"/>
         public TData JobData => this.GetJobData();
+
+#if UNITY_EDITOR
+        /// <summary>
+        /// Whether job data should be synchronized in the <c>OnValidate</c> message.
+        /// By default, returns true.
+        /// </summary>
+        /// <remarks>
+        /// Override in subclasses implementing <see cref="IJobDataSynchronizer{}"/> and return false
+        /// to avoid automatic job data synchronization during <c>OnValidate</c>.
+        /// </remarks>
+        protected virtual bool SynchronizeJobDataOnValidate => true;
+
+        /// <summary>
+        /// Synchronizes job data on Play mode if <see cref="SynchronizeJobDataOnValidate"/> returns true.
+        /// </summary>
+        protected virtual void OnValidate()
+        {
+            if (Application.isPlaying && SynchronizeJobDataOnValidate)
+            {
+                this.SynchronizeJobDataOnce();
+            }
+        }
+#endif
     }
 
     /// <summary>
     /// Alias for <see cref="AJobBehaviour{}"/>.
-    /// Pass <c>UpdateTransformJob&lt;<typeparamref name="TData"/>&gt;</c> as <typeparamref name="TJob"/> to Burst compile your job.
+    /// Pass <c>BurstUpdateTransformJob&lt;<typeparamref name="TData"/>&gt;</c> as <typeparamref name="TJob"/> to Burst compile your job.
     /// </summary>
     public abstract class AJobBehaviour<TData, TJob> : AJobBehaviour<TData>
         where TData : struct, IUpdateTransformJob