+++ Runtime/Jobs/UpdateJob.cs
@@ -2,10 +2,21 @@ using Gilzoide.UpdateManager.Jobs.Internal;
 
 namespace Gilzoide.UpdateManager.Jobs
 {
+    public struct UpdateJob<TData> : IInternalUpdateJob<TData>
+        where TData : struct, IUpdateJob
+    {
+        public UnsafeNativeList<TData> Data { get; set; }
+
+        public unsafe void Execute(int index)
+        {
+            Data.ItemRefAt(index).Execute();
+        }
+    }
+
 #if HAVE_BURST
     [Unity.Burst.BurstCompile]
 #endif
-    public struct UpdateJob<TData> : IInternalUpdateJob<TData>
+    public struct BurstUpdateJob<TData> : IInternalUpdateJob<TData>
         where TData : struct, IUpdateJob
     {
         public UnsafeNativeList<TData> Data { get; set; }